analyzer: move ChainID, Network, and Runtime out
