+++ analyzer/evmtokenbalances/evm_token_balances.go
@@ -79,7 +79,7 @@ const (
 )
 
 type Main struct {
-	runtime analyzer.Runtime
+	runtime common.Runtime
 	cfg     analyzer.RuntimeConfig
 	target  storage.TargetStorage
 	logger  *log.Logger
@@ -88,7 +88,7 @@ type Main struct {
 var _ analyzer.Analyzer = (*Main)(nil)
 
 func NewMain(
-	runtime analyzer.Runtime,
+	runtime common.Runtime,
 	sourceConfig *config.SourceConfig,
 	target storage.TargetStorage,
 	logger *log.Logger,
